requirements.txt
crewai[tools]
langchain-google-genai
python-dotenv
